+++ src/app/_shared/models/user.model.ts
@@ -3,22 +3,29 @@ import {Deserializable} from './deserializable.model';
 
 import * as moment from 'moment';
 
+export class Permission {
+  permissionName: string;
+}
+
 export class User implements Deserializable<User> {
 
   id: number;
+  firstName: string;
+  lastName: string;
   fullName: string;
   userName: string;
   email: string;
   roleID: number;
+  roleName: string;
   jobTitleID: number;
   jobSubTitleID: number;
-  loginEnabled: boolean;
-  forcePasswordReset: boolean;
   createdBy: number;
-  createdAt: string;
-  updatedBy: number;
-  updatedAt: string;
-  expiringAt: number;
+  creationDate: string;
+  lastUpdatedBy: number;
+  lastUpdateDate: string;
+  isManager: boolean;
+  isExistingUser: boolean;
+  permissions: Permission[];
 
   // TO-DO BILL: add comments for these three methods
 
@@ -27,12 +34,9 @@ export class User implements Deserializable<User> {
     return this;
   }
 
-  isLoginEnabled(): boolean {
-    return this.loginEnabled;
-  }
 
   minutesSinceLastUpdate(): number {
-    return moment().diff(moment(this.updatedAt), 'minutes');
+    return moment().diff(moment(this.lastUpdateDate), 'minutes');
   }
 
 }