+++ src/app/_shared/services/api-data/api-data-dashboard.service.ts
@@ -3,31 +3,38 @@ import { Http, Headers, Response, RequestOptions, ResponseContentType } from '@a
 import { Observable } from 'rxjs/observable';
 import { forkJoin } from 'rxjs/observable/forkJoin';
 import { CacheService } from '../cache.service';
+import { AuthService } from '../auth.service';
+import 'rxjs/add/operator/map';
 
 @Injectable()
 export class ApiDataDashboardService {
 
   constructor(
     private http: Http,
-    private cacheService: CacheService
+    private cacheService: CacheService,
+    private authService: AuthService
   ) { }
 
 
-  getDashboardData(emailAddress: string, startDate: string, endDate: string, userName: string, employeeID: number): Observable<any> {
+  getDashboardData(startDate: string, endDate: string): Observable<any> {
 
-    const fteData = this.http.get(`/api/dashboard/getFTEData/${emailAddress}/${startDate}/${endDate}`)
+    // NOTE: email is passed here instead of id as the key since it gets data from the plm databridge as well as jarvis
+    const headers = new Headers({'X-Token': this.cacheService.token.signedToken});
+    const options = new RequestOptions({headers: headers});
+
+    const fteData = this.http.get(`/api/dashboard/dashboard/show/getFTEData/${startDate}/${endDate}`, options)
       .timeout(this.cacheService.apiDataTimeout)
       .map((response: Response) => response.json());
 
-    const firstLogin = this.http.get(`/api/dashboard/checkFirstLogin/${employeeID}/${userName}`)
+    const firstLogin = this.http.get(`/api/dashboard/checkFirstLogin`, options)
       .timeout(this.cacheService.apiDataTimeout)
       .map((response: Response) => response.json());
 
-    const projectRequests = this.http.get(`/api/dashboard/checkProjectRequests/${employeeID}`)
+    const projectRequests = this.http.get(`/api/dashboard/checkProjectRequests`, options)
       .timeout(this.cacheService.apiDataTimeout)
       .map((response: Response) => response.json());
 
-    const jobTitle = this.http.get(`/api/dashboard/checkJobTitle/${employeeID}`)
+    const jobTitle = this.http.get(`/api/dashboard/checkJobTitle`, options)
       .timeout(this.cacheService.apiDataTimeout)
       .map((response: Response) => response.json());